combines catalog classes into one class
