+++ lib/catalog/abstract_adapter.rb
@@ -1,0 +1,9 @@
+class Catalog::AbstractAdapter
+  def load_catalog
+    raise "ERROR: Implementation Required"
+  end
+
+  def update_section_seats
+    raise "ERROR: Implementation Required"
+  end
+end
