--- lib/catalog/abstract_catalog_loader.rb
@@ -1,5 +1,0 @@
-class Catalog::AbstractCatalogLoader
-  def load_catalog
-    raise "ERROR: Implementation Required"
-  end
-end
