--- lib/catalog/abstract_catalog_updater.rb
@@ -1,5 +1,0 @@
-class Catalog::AbstractCatalogUpdater
-  def update_section_seats
-    raise "ERROR: Implementation Required"
-  end
-end
