+++ lib/tasks/catalog.rake
@@ -1,9 +1,9 @@
 namespace :catalog do
   task load: :environment do
-    Catalog::RpiCatalogLoader.new.load_catalog
+    Catalog::RpiAdapter.new.load_catalog
   end
 
   task update_seats: :environment do
-    Catalog::RpiCatalogUpdater.new.update_section_seats
+    Catalog::RpiAdapter.new.update_section_seats
   end
 end
